Improve shuffling of permutations
The shuffled version of get_permutations did not shuffle the data
sufficiently.  It simply shuffled the lists and passed them into
itertools.product producing a prodictable ordering.
To improve this a method was created to get the nth product and used
alongside a lazy shuffled range implementation to create a lazy
generator which produces a well shuffled output.